+++ forge/tx-util/lib/index.js
@@ -1,8 +1,4 @@
-/* eslint-disable object-curly-newline */
-const ForgeSDK = require('@arcblock/forge-sdk');
-const { EventEmitter } = require('events');
-
-const debug = require('debug')(require('../package.json').name);
+const { create, createAsync } = require('./verifier');
 
 /**
  * Generate a transaction verifier by tx hash
@@ -20,88 +16,82 @@ const debug = require('debug')(require('../package.json').name);
  * @param {number} params.maxRetry - max number of checks before mark the tx as expired
  * @returns {EventEmitter} verifier object
  */
-const createVerifier = params => {
-  const {
-    hash,
-    chainHost,
-    chainId = '',
-    checkInterval = 1000,
-    autoStart = true,
-    maxRetry = 20,
-  } = params;
-  debug('tx.verify.params', params);
-
-  if (!hash) {
-    throw new Error('Transaction hash must be provided to create verifier');
-  }
-  if (!chainHost) {
-    throw new Error('chainHost must be provided to create verifier');
-  }
-
-  ForgeSDK.connect(chainHost, { chainId, name: chainId || chainHost });
-
-  const events = new EventEmitter();
-  let retryCount = -1;
-
-  const doCheck = async () => {
-    if (retryCount > maxRetry) {
-      events.emit('error', {
-        hash,
-        type: 'expired',
-        retryCount,
-        error: new Error('Transaction verifier exceeded max retry count'),
-      });
-      return;
-    }
-
-    retryCount += 1;
-
-    try {
-      const res = await ForgeSDK.getTx({ hash }, { conn: chainId || chainHost });
-      if (res && res.code === 'OK' && res.info) {
-        if (res.info.code === 'OK') {
-          debug('tx.verify.done', { hash, retryCount });
-          events.emit('done', { hash, tx: res.info });
-        } else {
-          events.emit('error', {
-            hash,
-            type: 'invalid',
-            retryCount,
-            code: res.info.code,
-            error: new Error(res.info.code),
-          });
-        }
-      } else {
-        debug('tx.verify.pending', { hash, txCode: res.code, retryCount });
-        setTimeout(doCheck, checkInterval);
+const txVerifierParams = {
+  key: 'hash',
+  label: 'tx',
+  fn: 'getTx',
+  validate: res => {
+    if (res.info) {
+      if (res.info.code === 'OK') {
+        return true;
       }
-    } catch (err) {
-      // eslint-disable-next-line no-console
-      console.error(err);
-      setTimeout(doCheck, checkInterval);
-      debug('tx.verify.error', { hash, err, retryCount });
-      events.emit('error', {
-        hash,
-        type: 'exception',
-        error: err,
-        retryCount,
-      });
+
+      // If the tx failed, we throw an error
+      throw new Error(res.info.code);
     }
-  };
 
-  events.start = () => {
-    setTimeout(doCheck, checkInterval);
-  };
-
-  if (autoStart) {
-    setTimeout(() => {
-      events.start();
-    }, 0);
-  }
-
-  return events;
+    return false;
+  },
 };
+const verifyTx = create(txVerifierParams);
+const verifyTxAsync = createAsync(txVerifierParams);
+
+/**
+ * Generate a account verifier by address
+ * Emit events on account declared in a block
+ *
+ *  - `error`: when there is something wrong when verify the transaction
+ *  - `done`: when the transaction is successfully verified
+ *
+ * @param {object} params
+ * @param {string} params.address - account address to check
+ * @param {string} params.chainHost - on which chain to check the account
+ * @param {string} params.chainId - on which chain to check the account
+ * @param {number} params.checkInterval - query interval to see the account
+ * @param {boolean} params.autoStart - should the verifier start on create
+ * @param {number} params.maxRetry - max number of checks before mark the check as expired
+ * @returns {EventEmitter} verifier object
+ */
+const accountVerifierParams = {
+  key: 'address',
+  label: 'account',
+  fn: 'getAccountState',
+  validate: res => res.state,
+};
+const verifyAccount = create(accountVerifierParams);
+const verifyAccountAsync = createAsync(accountVerifierParams);
+
+/**
+ * Generate a asset verifier by address
+ * Emit events on asset can be found on chain
+ *
+ *  - `error`: when there is something wrong when verify the asset
+ *  - `done`: when the asset is found
+ *
+ * @param {object} params
+ * @param {string} params.address - asset address to check
+ * @param {string} params.chainHost - on which chain to check the asset
+ * @param {string} params.chainId - on which chain to check the asset
+ * @param {number} params.checkInterval - query interval to see the asset
+ * @param {boolean} params.autoStart - should the verifier start on create
+ * @param {number} params.maxRetry - max number of checks before mark the check as expired
+ * @returns {EventEmitter} verifier object
+ */
+const assetVerifierParams = {
+  key: 'address',
+  label: 'asset',
+  fn: 'getAssetState',
+  validate: res => res.state,
+};
+const verifyAsset = create(assetVerifierParams);
+const verifyAssetAsync = createAsync(assetVerifierParams);
 
 module.exports = {
-  createVerifier,
+  createVerifier: verifyTx,
+  verifyTx,
+  verifyTxAsync,
+  verifyAccount,
+  verifyAccountAsync,
+  verifyAsset,
+  verifyAssetAsync,
 };